Update URL regex to work with new launch pages
Zoom recently (within the last week) changed their launch pages such
that the URL matching in this extension no longer works. The new urls
are of the form:

  https://[company].zoom.us/j/1234567890#success

This PR updates the URL regex so that it works with both the old and new
style launch pages. Thanks for creating this extension!